Refine and generalise find by value visitor
* Fix unit tests
