+++ src/main/java/uk/gov/companieshouse/certificates/orders/api/controller/ApiErrors.java
@@ -23,7 +23,6 @@ public final class ApiErrors {
     private static final String GENERAL_PARTNER_DETAILS_SUPPLIED_ERROR = "general-partner-details-error";
     private static final String LIMITED_PARTNER_DETAILS_SUPPLIED_ERROR = "limited-partner-details-error";
     private static final String PRINCIPAL_PLACE_OF_BUSINESS_DETAILS_SUPPLIED_ERROR = "principal-place-of-business-details-error";
-    public static final String INCLUDE_DOB_TYPE_REQUIRED_ERROR = "include-dob-type-is-null-error";
     private static final String INCLUDE_EMAIL_COPY_NOT_ALLOWED_ERROR = "include-email-copy-is-true-error";
     private static final String COLLECTION_LOCATION_REQUIRED_ERROR = "collection-location-required-error";
     private static final String FORENAME_REQUIRED_ERROR = "forename-required-error";
@@ -120,7 +119,6 @@ public final class ApiErrors {
     public static final ApiError ERR_GENERAL_PARTNER_DETAILS_SUPPLIED = new ApiError(GENERAL_PARTNER_DETAILS_SUPPLIED_ERROR, GENERAL_PARTNER_DETAILS_LOCATION, OBJECT_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_LIMITED_PARTNER_DETAILS_SUPPLIED = new ApiError(LIMITED_PARTNER_DETAILS_SUPPLIED_ERROR, LIMITED_PARTNER_DETAILS_LOCATION, OBJECT_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_PRINCIPAL_PLACE_OF_BUSINESS_DETAILS_SUPPLIED = new ApiError(PRINCIPAL_PLACE_OF_BUSINESS_DETAILS_SUPPLIED_ERROR, PRINCIPAL_PLACE_OF_BUSINESS_DETAILS_LOCATION, OBJECT_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
-    public static final ApiError ERR_INCLUDE_DOB_TYPE_REQUIRED = new ApiError(INCLUDE_DOB_TYPE_REQUIRED_ERROR, INCLUDE_DOB_TYPE_LOCATION, OBJECT_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_INCLUDE_EMAIL_COPY_NOT_ALLOWED = new ApiError(INCLUDE_EMAIL_COPY_NOT_ALLOWED_ERROR, INCLUDE_EMAIL_COPY_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_COLLECTION_LOCATION_REQUIRED = new ApiError(COLLECTION_LOCATION_REQUIRED_ERROR, COLLECTION_LOCATION_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_FORENAME_REQUIRED = new ApiError(FORENAME_REQUIRED_ERROR, FORENAME_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
@@ -143,13 +141,13 @@ public final class ApiErrors {
     public static final ApiError ERR_DIRECTOR_DETAILS_INCLUDE_COUNTRY_OF_RESIDENCE_ERROR = new ApiError(INCLUDE_COUNTRY_OF_RESIDENCE_ERROR, DIRECTOR_DETAILS_INCLUDE_COUNTRY_OF_RESIDENCE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_DIRECTOR_DETAILS_INCLUDE_NATIONALITY = new ApiError(INCLUDE_NATIONALITY_ERROR, DIRECTOR_DETAILS_INCLUDE_NATIONALITY_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_DIRECTOR_DETAILS_INCLUDE_OCCUPATION = new ApiError(INCLUDE_OCCUPATION_ERROR, DIRECTOR_DETAILS_INCLUDE_OCCUPATION_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
-    public static final ApiError ERR_DIRECTOR_DETAILS_INCLUDE_DOB_TYPE_SUPPLIED_ERROR = new ApiError(INCLUDE_DOB_TYPE_REQUIRED_ERROR, DIRECTOR_DETAILS_INCLUDE_DOB_TYPE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
+    public static final ApiError ERR_DIRECTOR_DETAILS_INCLUDE_DOB_TYPE_SUPPLIED_ERROR = new ApiError(INCLUDE_DOB_TYPE_ERROR, DIRECTOR_DETAILS_INCLUDE_DOB_TYPE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_SECRETARY_DETAILS_INCLUDE_ADDRESS = new ApiError(INCLUDE_ADDRESS_ERROR, SECRETARY_DETAILS_INCLUDE_ADDRESS_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_SECRETARY_DETAILS_INCLUDE_APPOINTMENT_DATE = new ApiError(INCLUDE_APPOINTMENT_DATE_ERROR, SECRETARY_DETAILS_INCLUDE_APPOINTMENT_DATE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_SECRETARY_DETAILS_INCLUDE_COUNTRY_OF_RESIDENCE_ERROR = new ApiError(INCLUDE_COUNTRY_OF_RESIDENCE_ERROR, SECRETARY_DETAILS_INCLUDE_COUNTRY_OF_RESIDENCE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_SECRETARY_DETAILS_INCLUDE_NATIONALITY = new ApiError(INCLUDE_NATIONALITY_ERROR, SECRETARY_DETAILS_INCLUDE_NATIONALITY_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_SECRETARY_DETAILS_INCLUDE_OCCUPATION = new ApiError(INCLUDE_OCCUPATION_ERROR, SECRETARY_DETAILS_INCLUDE_OCCUPATION_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
-    public static final ApiError ERR_SECRETARY_DETAILS_INCLUDE_DOB_TYPE_SUPPLIED_ERROR = new ApiError(INCLUDE_DOB_TYPE_REQUIRED_ERROR, SECRETARY_DETAILS_INCLUDE_DOB_TYPE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
+    public static final ApiError ERR_SECRETARY_DETAILS_INCLUDE_DOB_TYPE_SUPPLIED_ERROR = new ApiError(INCLUDE_DOB_TYPE_ERROR, SECRETARY_DETAILS_INCLUDE_DOB_TYPE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
 
     private ApiErrors() {}
