+++ src/main/java/uk/gov/companieshouse/certificates/orders/api/model/DesignatedMemberDetails.java
@@ -16,7 +16,7 @@ public class DesignatedMemberDetails implements DateOfBirthIncludable<Map<String
     private Boolean includeBasicInformation;
     private Boolean includeCountryOfResidence;
     private IncludeDobType includeDobType;
-    private transient final Map<String, Object> fieldValues = new HashMap<>();
+    private final Map<String, Object> fieldValues = new HashMap<>();
 
     public Boolean getIncludeAddress() {
         return includeAddress;
